python/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
waitress==2.1.2
python-dotenv==1.0.0
Flask-Login==0.6.2
Werkzeug==2.3.7